cmd/gorilla-ws-chat: Initial, gorilla websocket-based implementation
Initial implementation of a chat server. This is still a hack-y test, as
it doesn't handle any error and just stop goroutines if anything fails
(without even clearing resources).

Connections to the server must at least have 2 parts, the last of which is
the username and the remaining identifies the chat room.