go_chat_i_guess: Add the chat server access token
Initial implementation of a chat server.

This chat server should be called from a web server to initialize a chat
channel (essentially, a room) and connect clients through a websocket
connection. Authentication is delegated to the web server, which must
request a token for authenticated and authorized users before connecting
the user to a chat channel.

Currently, the server is only able to generate access token for users, but
it still cannot consume those tokens from a public interface.