go_chat_i_guess server: Add channels to the chat server
Improve the chat server interface with capabilities to create and retrieve
channels.

A channel is automatically removed from the chat server after it gets
closed, regardless if because of a timeout or because it was manually
closed.

Channels are uniquely identified from their names, which may be defined by
the calling application. Alternativelly, an application could generate a
(pseudo) random name and until `CreateChannel()` returned `nil.