go_chat_i_guess: Refactor stopping channel
Some structures may wait indefinitely on one or more channels. When these
structures are close, an accompanying channel is used to avoid stalling
until the any of main channels receive a message.

Originally, a boolean would be pushed into these accompanying channels when
the loop was supposed to exit. However, closing a channel causes every read
from that channel to return immediately.

So, refactor these channels into a channel of struct{} and use the closed
property of the channel to signal that the loop should stop.