go_chat_i_guess: Implement connecting users
Implement connecting a user to a channel in a connection-agnostic way.

The "external facing server" (be it a HTTP server that will use WebSockets
for the chat, a TCP server or whatever) must maintain a stateful connection
to users, that implements this package's `Conn` interface. Internally to
the Chat Server, this connection gets associated with a specific username
in a specific channel.