cmd/chat-server: Add the chat endpoint(s)
Add every endpoint required to access the chat, and update the page to
properly use the Chat Server. The new endpoints are:

* /new_channel
* /new_token
* /chat

To access the chat, first a new channel must be created by accessing
`/new_channel/<channel_name>`. Then, each user must request a token in the
`/new_token/<channel_name>/<username>`, which is returned in the body of
the response. Finally, `/chat` must be accessed and upgraded to a
WebSocket, passing the token as the cookie `X-ChatToken`.

This implementation completely ignores authentication! Ideally, the server
should possibly restrict in some way who is able to create channels, and
must authenticated if the person requesting the token does actually own
that username.