go_chat_i_guess server: Connect the user on the current goroutine
Add function `ConnectAndWait` to the `ChatServer` so users may use the
current goroutine, instead of always spawning a new goroutine. This may be
useful in servers (e.g., web server) that already spawn a new goroutine to
handle new connections.

As an example, `cmd/chat-server` now handles the connected user in the same
goroutine that handles the received HTTP request.