cmd/chat-server: Fix SIGSEGV after a client disconnects
A client disconnection may be sent to the client that just disconnected.
Because of how lazily the example Conn had been write, this would cause an
attempt to `Write()` to a closed WebSocket connection, resulting in a
segmentation fault.

Check if the connection is null to fix segfaulting when a user disconnects.