go_chat_i_guess channel: Add custom encoding of messages
Add interface `MessageEncoder` which may optionally be defined when
creating a `ChatServer` to encode and/or process received messages.

Encoding the message may be as simples as concatenating every field, or the
message could be encoded into a JSON object. Additionally, the encoder may
parse the received message and process it, executing commands in the server
(for example, listing the users in the channel).

The example application `cmd/chat-server` simply returns the message as the
default encoder would, but it now parses the message "/users" and returns
the list of connected users to whoever sent the message.