go_chat_i_guess channel: Refactor handling an idle channel
After the first user connects to a channel, the channel would only detect
an idle or disconnected `Conn` when it received a message. So, unless the
`Conn` actively sent a message before closing, a channel could stay idle
forever.

Refactor the channel's event loop to wait for a message or for a timeout
event. In case of a timeout, the channel sends an empty message to every
user to check whether they `Conn` is still active.

This change causes `go_chat_i_guess` to require Go 1.15 or more recent,
since the timeout uses `time.Ticker.Reset()`.