go_chat_i_guess channel: Add "OnEnter" and "OnExit" events
Refactor the message controller so an application may define custom
behavior when a user connects, on the `OnConnect` event, or disconnects, on
the `OnDisconnect` event.

Since the disconnection happens while the channel map is locked, only a few
channel functions may be called from that event. To limit that, part of
`ChatChannel` was refactored into a super-interface,
`RestrictedChatChannel`,  which has access to all thread-safe functions
available to the channel.

To keep all channel processing into a single object, `ChannelController`
was defined as a sub-interface of `MessageEncoder`. Whenever a new channel
is created, the server checks whether the supplied `Controller` implements
both interfaces and, in that case, it sets the channel's `controller`.